+++ src/test/java/ma/vi/esql/parse/ParseTest.java
@@ -6,7 +6,6 @@ package ma.vi.esql.parse;
 
 import ma.vi.esql.DataTest;
 import ma.vi.esql.exec.EsqlConnection;
-import ma.vi.esql.exec.Result;
 import ma.vi.esql.parser.Parser;
 import ma.vi.esql.parser.query.Select;
 import org.junit.jupiter.api.DynamicTest;